refactor: move Helm package to internal
